requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Tracking
supervision>=0.16.0

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0

# Performance
scipy>=1.10.0

# Optional: GPU acceleration
# torch>=2.0.0+cu118  # Uncomment for CUDA 11.8
# torchvision>=0.15.0+cu118

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
